+++ tests/conftest.py
@@ -125,12 +125,13 @@ def store_with_events(store_factory):
 
 @pytest.fixture
 def stream_factory(identifier, db_engine):
-    def _inner(store: MessageStore[AccountEvent], partitioner=None):
+    def _inner(store: MessageStore[AccountEvent], partitioner=None, batch_size=None):
         stream = AggregatedStream[AccountEvent](
             name=identifier(),
             store=store,
             partitioner=partitioner or MyPartitioner(),
             stream_wildcards=["account-%"],
+            update_batch_size=batch_size,
         )
         with db_engine.connect() as conn:
             stream.truncate(conn)