+++ docs/CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.11.0 (only release candidate)
 
 * Improved performance of aggregated stream projection.
+* Fixed bug in return of AggregatedStream.projector.run which would always re-run it in the executor right away.
 * **BREAKING CHANGE**: Extended the RunOnNotification protocol
     * Added `interested_in_notification(dict) -> bool`
     * Added `take_notification_hint(dict) -> bool`